requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.8
django-allauth==0.45.0
django-cors-headers==3.10.0
django-filter==21.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.2
mysqlclient==2.0.3
oauthlib==3.1.1
pycparser==2.20
PyJWT==2.2.0
PyMySQL==1.0.2
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
urllib3==1.26.7
whitenoise==5.3.0
dj-rest-auth==2.1.11
django-rest-auth==0.9.5
